moved client/servery proxies back to main cornucopia package
